fix(server): resolve pyrefly type errors across server modules and tests
- Narrow download directory types in server/app.py and server/config.py before filesystem operations
- Fix safe title lowercasing and title narrowing in server/downloader.py, removing unused variables
- Annotate build_ydl_options return type as dict[str, Any] and cast ydl_opts to Any for YoutubeDL constructor
- Add runtime type guards for filename tracking and info dict extractions in server/engine.py
- Document dynamic monkey-patching in server/patches.py with targeted, explained pyrefly ignore comments
- Modernize format_eta parameter annotation to float | None per PEP 484 and PYI041
- Narrow MockYoutubeDL.last_instance in tests/server/test_mock_downloader.py with explicit non-null assertions
- Add documented pyrefly ignore comments in tests/server/test_remote_control_php.py for monkey-patched determine_ext

+++ server/utils.py
@@ -14,7 +14,7 @@ def format_bytes(b: float | None) -> str:
     return f"{b:.1f} PB"
 
 
-def format_eta(seconds: int | None) -> str:
+def format_eta(seconds: float | None) -> str:
     """Formats remaining seconds into mm:ss or hh:mm:ss format."""
     if seconds is None or seconds < 0:
         return "--:--"
